refactor(cli): move setup wizard into commands/setup/ submodule
Split 1948-line setup.rs into 8 focused modules:
- mod.rs: wizard loop and action dispatch
- provider.rs: LLM provider configuration and OAuth
- agent.rs: agent configuration
- channel.rs: channel + routing configuration
- tools.rs: tool configuration
- config_io.rs: shared I/O utilities
- scan.rs: config state scanner (was setup_scan.rs)
- ui.rs: terminal rendering (was setup_ui.rs)
