+++ crates/clawhive-cli/src/commands/setup/scan.rs
